+++ src/components/birthdays/BirthdaysManager.tsx
@@ -1,3 +1,0 @@
-import FamilyEventsManager from "@/components/birthdays/FamilyEventsManager";
-
-export default FamilyEventsManager;